==> code/04_HR_data_processing.Rmd <==
---
title: "Effects of temperature on Califonria spiny lobster (*Panulirus interruptus*) heart rates"
author: "Sam Csik"
date: "7/18/2019"
output: html_document
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

###**Summary**

Star-Oddi [DST milli-HRT loggers](https://www.star-oddi.com/products/archival-tags/heart-rate-monitor-animals) were surgically implanted into experimental lobsters following the completion of all respriometry and functional response trials to compare heart rates across treatments. This idea was originally proposed because of anomolous readings captured during the 11C lobster respirometry trials--it appeared as though lobsters were not breathing for durations of time exceeding 10 minutes repeatedly throughout the 24 hour-long trial.

We successfully captured ~24hr of heart rate data for n = 3 lobsters per treatment, except for 26C lobsters (due to equipment malfunctions).

To prepare data for analysis, we visually assessed electrocardiogram (ECG) traces to ensure that quality heart rates were estimated. Here, we wrangle the raw data files and print out ECG traces and heart rates (bpm) through time for visual assessment. We found that any erroneous recordings were often associated with two types of anomolous data: (1) In some cases, ECG traces appear irregular and corresponded to atypical heart rate recordings. This is presumably from insufficient contact between the logger's electrodes and the cardiac tissue. These recordings are noted in the `poor_signal_inventory.csv` file and subsequently removed from the data set. (2) Alternatively, arrested cardiac function was evident from a cessation in R wave signals in some ECGs. These instances were always associated with biologically unrealistic heart rates (bpm > 250). These recordings were noted in the `hr_stop_inventory.csv` file and subsequently replaced with "0 bpm."

###**Outline**

**Part 0** Import packages and metadata
**Part 1** Import and wrangle all beats per minute (bpm) and electrocardiogram (ecg) files
**Part 2** Combine all raw files into a single master file
**Part 3** Plot bpm data for each lobster to determine which are usuable; color code measurements by QI value (0 = good; 3 = bad)
**Part 4** Plot ecg data in 6s windows (our chosen measurement frequency)
**Part 5** Automated data processing to remove poor quality measurements and ensure that periods corresponding to arrested heart rates are recorded as "0 bpm"

*NOTE: ECGs generated in Part 3 were visually assessed to determine if measurement is truly reliable (i.e. produced clear QRS pulses). If strange/inconsistent signals were noted, the corresponding bpm measurements were noted in the `poor_signal_inventory.csv` file for removal. Similarly, some lobsters had arrested heart rates, which were evident from absent R-wave signals. These measurements always corresponded to biologially-unrealistic heart rates and were noted in the `hr_stop_inventory.csv` file for replacement with "0."
 
###**Packages Required:** 

    - tidyverse
    - here
    
###**Data Required:**

**(file path: data/metadata)**

    - lobster_heart_rate.csv
    
**(file path: data/heart_rate/raw/bpm)**

    - 9UHL0525_IV10_cold_DAT.csv
    - 1UHL0530_IV11_cold_DAT.csv
    - 5UHL0526_IV12_cold_DAT.csv
    - 6UHL0528_IV13_cold_DAT.csv
    - 1UHL0529_IV14_cold_DAT.csv
    - 6UHL0527_MM02_cold_DAT.csv
    - 4UHL0529_GP05_mid_DAT.csv
    - 9UHL0528_GP06_mid_DAT.csv
    - 8UHL0526_IV16_mid_DAT.csv
    - 10UHL0527_IV17_mid_DAT.csv
    - 4UHL0530_IV18_mid_DAT.csv
    - 7UHL0528_L3_hot_DAT.csv
    - 8UHL0528_L4_hot_DAT.csv
    - UHL0526_N14_hot_DAT.csv
    - 2UHL0530_N15_hot_DAT.csv
    - 2UHL0529_N19_hot_DAT.csv
    - 12UHL0525_GP01_extrahot_DAT.csv
    - 9UHL0527_GP07_extrahot_DAT.csv
    - 7UHL0526_IV19_extrahot_DAT.csv
    
**(file path: data/heart_rate/raw/ecg)**

    - 9UHL0525_IV10_cold_HAT.csv
    - 1UHL0530_IV11_cold_HATcsv
    - 5UHL0526_IV12_cold_HAT.csv
    - 6UHL0528_IV13_cold_HAT.csv
    - 1UHL0529_IV14_cold_HAT.csv
    - 6UHL0527_MM02_cold_HAT.csv
    - 4UHL0529_GP05_mid_HAT.csv
    - 9UHL0528_GP06_mid_HAT.csv
    - 8UHL0526_IV16_mid_HAT.csv
    - 10UHL0527_IV17_mid_HAT.csv
    - 4UHL0530_IV18_mid_HAT.csv
    - 7UHL0528_L3_hot_HAT.csv
    - 8UHL0528_L4_hot_HAT.csv
    - UHL0526_N14_hot_HAT.csv
    - 2UHL0530_N15_hot_HAT.csv
    - 2UHL0529_N19_hot_HAT.csv
    - 12UHL0525_GP01_extrahot_HAT.csv
    - 9UHL0527_GP07_extrahot_HAT.csv
    - 7UHL0526_IV19_extrahot_HAT.csv
    
**(file path: data/heart_rate/inventory_files)**

    - poor_signal_inventory.csv
    - hr_stop_inventory.csv
    
###**Part 0.** Load packages & data 
```{r, message = FALSE, results = 'hide', warning = FALSE}
##############################
# load required packages
##############################

source(here::here("code", "00_libraries.R")) 

##############################
# import metadata
##############################

# used in for loops to import all data files
heart_rate_metadata <- read_csv(here::here("data", "metadata", "metadata_heart_rate.csv"))

# used for processing the data
poor_signal_inventory <- read_csv(here::here("data", "heart_rate", "inventory_files", "poor_signal_inventory.csv")) %>% 
  filter(remove == "yes")

# used for processing the data
hr_stop_inventory <- read_csv(here::here("data", "heart_rate", "inventory_files", "hr_stop_inventory.csv")) %>% 
  filter(alter_bpm == "yes")
```

###**Part 1.** Import all HR data (beats per min) and wrangle data using `import_bpm()` function
```{r}
##############################
# function reads in each bpm data file, change header names, add lobster ID and temp to df
##############################

import_bpm <- function(file_name, lobster_id, temperature){
  read_csv(here::here("data/heart_rate/raw/bpm", file_name)) %>% 
  rename(date_time = X, temp = `Temp(C)`, hr = `Heart rate(bpm)`) %>% 
  select(date_time, temp, hr, QI) %>% 
  mutate(lobster_id = rep(lobster_id), 
         temperature = rep(temperature))
}

##############################
# function reads in each ecg data file
##############################

import_ecg <- function(file_name, lobster_id, temperature){
  read_csv(here::here("data/heart_rate/raw/ecg", file_name)) %>% 
  rename(date_time = X, ecg = `Bin ECG`) %>% 
  mutate(lobster_id = rep(lobster_id), 
         temperature = rep(temperature))
}

##############################
# for loop automates reading in each bpm and ecg file using import_bpm and import_ecg functions
##############################

# DAT (bpm) files
for (row in 1:nrow(heart_rate_metadata)) {
  data_row <- heart_rate_metadata[row,]
  file_name <- as.character(data_row[, 1]) 
  lobster_id <- as.character(data_row[, 3]) 
  temperature <- as.character(data_row[, 4]) 
  save_name <- paste(lobster_id, "bpm", sep = "_")
  table <- import_bpm(file_name, lobster_id, temperature) 
  assign(save_name, table) 
}

# HAT (ecg files)
for (row in 1:nrow(heart_rate_metadata)) {
  data_row <- heart_rate_metadata[row,]
  file_name <- as.character(data_row[, 2])
  lobster_id <- as.character(data_row[, 3])
  temperature <- as.character(data_row[, 4])
  save_name <- paste(lobster_id, "ecg", sep = "_")
  table <- import_ecg(file_name, lobster_id, temperature) 
  assign(save_name, table) 
}
```

####**Part 2.** Combine all lobster files
```{r}
##############################
# combine all lobster bpm dfs
##############################

all_bpm <- rbind(IV10_bpm, IV11_bpm, IV12_bpm, IV13_bpm, IV14_bpm, MM02_bpm, 
                GP05_bpm, GP06_bpm, IV16_bpm, IV17_bpm, IV18_bpm, 
                L3_bpm, L4_bpm, N14_bpm, N15_bpm, N19_bpm, 
                GP01_bpm, GP07_bpm, IV19_bpm) %>% 
  mutate(temperature = as.factor(temperature),
         lobster_id = fct_relevel(lobster_id, 
                                  "IV10", "IV11", "IV12", "IV13", "IV14","MM02",
                                  "GP05", "GP06", "IV16", "IV17", "IV18",
                                  "L3", "L4", "N14", "N15", "N19",
                                  "GP01", "GP07", "IV19")) %>% 
  mutate(date2 = mdy_hm(date_time))

##############################
# combine all lobster ecg dfs
##############################

all_ecg <- rbind(IV10_ecg, IV11_ecg, IV12_ecg, IV13_ecg, IV14_ecg, MM02_ecg, 
                GP05_ecg, GP06_ecg, IV16_ecg, IV17_ecg, IV18_ecg, 
                L3_ecg, L4_ecg, N14_ecg, N15_ecg, N19_ecg, 
                GP01_ecg, GP07_ecg, IV19_ecg) %>% 
  mutate(lobster_id = fct_relevel(lobster_id, 
                                  "IV10", "IV11", "IV12", "IV13", "IV14","MM02",
                                  "GP05", "GP06", "IV16", "IV17", "IV18",
                                  "L3", "L4", "N14", "N15", "N19",
                                  "GP01", "GP07", "IV19")) %>% 
  mutate(date2 = mdy_hm(date_time))
```

---------------------------------
---------------------------------
Visualize raw bpm and ecg data
---------------------------------
---------------------------------

####**Part 3.** Plot HR ~ time for each individual; color code each measurement by it's assigned QI value
```{r}
##############################
# define unique values for each lobster_id
##############################

IDs <- unique(heart_rate_metadata$lobster_id) 

##############################
# for loop to plot hr through time for each lobster; color code by QI value
##############################

# pdf(paste("figures/other/bpm_assessment",".pdf", sep = "_"), onefile = TRUE, width = 9, height = 6)

for(i in 1:length(IDs)){ # set range of 1 to the length of IDs
  
  lob_id <- IDs[i] # assign IDs[index] to new var lob_id so it makes more intuitive sense
  
  subset <- all_bpm %>% # subset the current index from all_hr
    filter(lobster_id == lob_id) 
    #filter(! QI %in% c("2", "3"))

  plot <- ggplot(subset, aes(x = date2, y = hr)) + 
    geom_point(aes(color = as.factor(QI))) + 
    labs(x = "Time", y = "Heart Rate (bpm)") +
    ggtitle(as.character(lob_id))
 
  print(plot)

}

# dev.off()
```

###**Part 4.** Plot ECGs by 6 sec interval windows for each lobster
```{r}
#############################
# function to break apart ECGs into 6 sec intervals and plot
##############################

make_ecg_pdf <- function(lob_id){
  # find the data for given lobster_id
  subset <- all_ecg %>% 
    filter(lobster_id == lob_id) 
  
  # find all unique date times from lobster data
  data_times <- unique(subset$date_time)
  
  # for each unique time, create a plot
  for(i in 1:length(data_times)){
    time_id <- data_times[i]
    filtered_times <- subset %>% 
      filter(date_time == time_id) %>% 
      mutate(id = row_number())
    
    # get title for plot 
    date_time <- filtered_times[1,1]
    
    # make the plot
    ecg_plot <- ggplot(filtered_times, aes(x = id, y = ecg)) + 
      geom_line() + 
      scale_x_continuous(breaks = seq(0, 600, by = 50)) +
      labs(x = "value", y = "ECG") +
      ggtitle(as.character(date_time))
    
    print(ecg_plot)
  }
}

#############################
# for loop makes ecg pdfs (for each 6sec interval) for each lobster
##############################

lob_ids <- unique(all_ecg$lobster_id)

for(i in 1:length(lob_ids)) {
  lob_id = lob_ids[i] # for each lob...
  pdf(paste("figures", lob_id, "ecg",".pdf", sep = "_"), onefile = TRUE, width = 9, height = 6)
  make_ecg_pdf(lob_id)
  dev.off()
}
```

---------------------------------
---------------------------------
Process data
---------------------------------
---------------------------------

####**Part 5.** Remove anomolous measurements (noted in `poor_signal_inventory.csv`); replace clear instances of arrested heart rates with "0" (noted in `hr_stop_inventory.csv`); select only usable lobsters
```{r}
##############################
# post process bpm data
##############################

# set counter to 0
clean_count <- 0

# build empty df
new_df <- data.frame(lobster_id = as.character(),
                     time = as.character(),
                     remove = as.character())

# for loop to replace erroneus hr values with manually assigned values from hr_inventory.csv and set values that should be removed completely to "NA"
for (i in 1:nrow(all_bpm)) {
  current_row <- all_bpm[i,]
  ts <- as.character(current_row[1][[1]])
  lob_id <- as.character(current_row[5][[1]])
  
  # subset replacement values
  hr_subset <- hr_stop_inventory %>% 
    filter(lobster_id == lob_id) %>% 
    filter(hr_stop_time == ts)
  
  # subset removal values
  poor_signal_subset <- poor_signal_inventory %>% 
    filter(lobster_id == lob_id) %>% 
    filter(time == ts)
   
  # if there are values to be replaced in all_bpm "hr" column, replace with value in hr_subset "alter_bpm" column
  if (nrow(hr_subset) > 0) {
    all_bpm[i,][3] <- hr_subset[5]
  }
  
  # if there are values to be removed, enter "NA" in those rows
  if (nrow(poor_signal_subset) > 0) {
    # clean_count should be equal to the number of rows in poor_signal_inventory
    print("----------------------")
    print("hey i'm removing stuff")
    print(all_bpm[i,])
    clean_count <- clean_count + 1
    all_bpm[i,][1] <- NA
    all_bpm[i,][2] <- NA
    all_bpm[i,][3] <- NA
    all_bpm[i,][4] <- NA
    all_bpm[i,][5] <- NA
    all_bpm[i,][6] <- NA
    new_df <- rbind(new_df, poor_signal_subset)
  }
  
  # sanity check
  if (i == nrow(all_bpm)) {
    print("Is My Post Processing Done?")
    print(clean_count == nrow(poor_signal_inventory))
  }
}

# print any rows that are missing from poor_signal_inventory
for (i in 1:nrow(poor_signal_inventory)) {
  current_row <- poor_signal_inventory[i,]
  ts <- as.character(current_row[2][[1]])
  lob_id <- as.character(current_row[1][[1]])
  # subset removal values
  sub <- new_df %>% 
    filter(lobster_id == lob_id) %>% 
    filter(time == ts)
  
  if (nrow(sub) == 0) {
    print(lob_id)
    print(ts)
  }
}

##############################
# I'm only going to keep lobs: 
  # 11C: IV11, IV13, IV14
  # 16C: GP05, GP06, IV17
  # 21C: L3, L4, N14
# filter out all other lobs from all_bpm
##############################

keepers <- all_bpm %>% 
  drop_na(lobster_id) %>% 
  mutate(lobster_id = as.factor(lobster_id),
         hr = as.numeric(hr),
         temperature = as.factor(temperature))  %>% 
  filter(lobster_id %in% c("IV11", "IV13", "IV14", "GP05", "GP06", "IV17", "L3", "L4", "N14"))
 
##############################
# filter out any extra values that are unreasonable
  # 11C > 80bpm
  # 16C > 110bpm
  # 21C > 130bpm
##############################

filtered11 <- keepers %>% 
  filter(temperature == "11",
         hr < 80)

filtered16 <- keepers %>% 
  filter(temperature == "16",
         hr < 110)

filtered21 <- keepers %>% 
  filter(temperature == "21",
         hr < 130)

processed_data <- rbind(filtered11, filtered16, filtered21) %>% 
  mutate(hr = as.numeric(hr),
         temp = as.numeric(temp)) %>% 
  rename(date = date2) %>% 
  select(date, temperature, lobster_id, hr, QI) %>% 
  rename(temp = temperature)

# write.csv(processed_data, here::here("data", "heart_rate", "outputs", "processed_heart_rates.csv"), row.names = FALSE)
```
