==> code/6_MR_HR_plotting.Rmd <==
---
title: "Plot metabolic rates and heart rates"
author: "Sam Csik"
date: "8/1/2019"
output: html_document
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

###**Summary**

Create plot of lobster SMR, MMR, AAS, FAS, and resting heart rate where individual lobster measurements are in greyscale and treatment means are colored points (with 95% CI).

###**Outline**

**Part 0** Import packages and data
**Part 1** Create plots of SMR, MMR, AAS, FAS
**Part 2** Create plots of resting heart rate
**Part 3** (FIGURE 2 GENERATED HERE) Combine plots and print

###**Packages Required:** 

    - tidyverse
    - here
    - cowplot
    
###**Data Required** 

**(file path: data/metabolism):**

    - metabolic_traits.csv

**(file path: data/metabolism/outputs):**

    - smr_summary_stats.csv
    - mmr_summary_stats.csv
    - aas_summary_stats.csv
    - fas_summary_stats.csv

**(file path: data/heart_rate/outputs):**

    - resting_heart_rates.csv
    - rhr_summary_stats.csv

####**Part 0.** Import libraries & metadata
```{r}
##############################
# import libraries
##############################

source(here::here("code", "0_libraries.R"))

##############################
# load data
##############################

metabolic_traits <- read_csv(here::here("data", "metabolism", "metabolic_traits.csv")) %>%
  mutate(temp = as.factor(temp))

resting_heart_rates <- read_csv(here::here("data", "heart_rate", "outputs", "resting_heart_rates.csv"))

smr_sum <- read_csv(here::here("data", "metabolism", "outputs", "smr_summary_stats.csv"))
mmr_sum <- read_csv(here::here("data", "metabolism", "outputs", "mmr_summary_stats.csv"))
aas_sum <- read_csv(here::here("data", "metabolism", "outputs", "aas_summary_stats.csv"))
fas_sum <- read_csv(here::here("data", "metabolism", "outputs", "fas_summary_stats.csv"))
rhr_sum <- read_csv(here::here("data", "heart_rate", "outputs", "rhr_summary_stats.csv"))
```

####**Part 2.** Plot SMR, MMR, AAS of lobsters in each temperature treatment using the q~15~ SMR method
```{r}
##############################
# smr panel
##############################

smr_plot <- ggplot() +
  geom_jitter(data = metabolic_traits, aes(x = as.factor(temp), y = SMR), color = "black", width = 0.2, alpha = 0.2, size = 3) +
  geom_errorbar(data = smr_sum, aes(x = as.factor(temp), ymin = SMR_avg - error, ymax = SMR_avg + error), width = 0.1, size = 1, color = "black") +
  geom_point(data = smr_sum, aes(x = as.factor(temp),  y = SMR_avg), size = 6, colour = "black") +
  geom_point(data = smr_sum, aes(x = as.factor(temp),  y = SMR_avg, color = as.factor(temp)), size = 4.5) +
  labs(y = expression(atop("Standard Metabolic Rate", paste("(",mg~O[2]~kg^-1~min^-1,")"))),
       x = "Temperature (°C)") +
  scale_y_continuous(breaks = seq(0, 1.6, by = 0.5)) +
  scale_color_manual(values=c("lightslategray", "lightblue", "lightcoral", "indianred4")) +
  theme_classic() +
  theme(axis.text = element_text(color = "black", size = 12),
        axis.title = element_text(size = 13),
        panel.border = element_rect(colour = "black", fill = NA, size = 0.7),
        plot.caption = element_text(size = 10, hjust = 0),
        legend.position = "none")

##############################
# mmr panel
##############################

mmr_plot <- ggplot() +
  geom_jitter(data = metabolic_traits, aes(x = as.factor(temp), y = MMR), color = "black", width = 0.2, alpha = 0.2, size = 3) +
  geom_errorbar(data = mmr_sum, aes(x = as.factor(temp), ymin = MMR_avg - error, ymax = MMR_avg + error), width = 0.1, size = 1, color = "black") +
  geom_point(data = mmr_sum, aes(x = as.factor(temp),  y = MMR_avg), size = 6, color = "black") +
  geom_point(data = mmr_sum, aes(x = as.factor(temp),  y = MMR_avg, color = as.factor(temp)), size = 4.5) +
  labs(y = expression(atop("Maximum Metabolic Rate", paste("(",mg~O[2]~kg^-1~min^-1,")"))),
       x = "Temperature (°C)") +
  ylim(0, 6.5) +
  scale_color_manual(values=c("lightslategray", "lightblue", "lightcoral", "indianred4")) +
  theme_classic() +
  theme(axis.text = element_text(color = "black", size = 12),
        axis.title = element_text(size = 13),
        panel.border = element_rect(colour = "black", fill = NA, size = 0.7),
        plot.caption = element_text(size = 10, hjust = 0),
        legend.position = "none")

##############################
# aas panel
##############################

aas_plot <- ggplot() +
  geom_jitter(data = metabolic_traits, aes(x = as.factor(temp), y = AAS), color = "black", width = 0.2, alpha = 0.2, size = 3) +
  geom_errorbar(data = aas_sum, aes(x = as.factor(temp), ymin = AS_avg - error, ymax = AS_avg + error), width = 0.1, size = 1, color = "black") +
  geom_point(data = aas_sum, aes(x = as.factor(temp),  y = AS_avg), size = 6, color = "black") +
  geom_point(data = aas_sum, aes(x = as.factor(temp),  y = AS_avg, color = as.factor(temp)), size = 4.5) +
  labs(y = expression(atop("Absolute Aerobic Scope", paste("(",mg~O[2]~kg^-1~min^-1,")"))),
       x = "Temperature (°C)") +
  ylim(0, 6.5) +
  scale_color_manual(values = c("lightslategray", "lightblue", "lightcoral", "indianred4")) +
  theme_classic() +
  theme(axis.text = element_text(color = "black", size = 12),
        axis.title = element_text(size = 13),
        panel.border = element_rect(colour = "black", fill = NA, size = 0.7),
        plot.caption = element_text(size = 10, hjust = 0),
        legend.position = "none")

##############################
# fas panel
##############################

fas_plot <- ggplot() +
  geom_jitter(data = metabolic_traits, aes(x = as.factor(temp), y = FAS), color = "black", width = 0.2, alpha = 0.2, size = 3) +
  geom_errorbar(data = fas_sum, aes(x = as.factor(temp), ymin = FAS_avg - error, ymax = FAS_avg + error), width = 0.1, size = 1, color = "black") +
  geom_point(data = fas_sum, aes(x = as.factor(temp),  y = FAS_avg), size = 6, color = "black") +
  geom_point(data = fas_sum, aes(x = as.factor(temp),  y = FAS_avg, color = as.factor(temp)), size = 4.5) +
  labs(y = expression(atop("Factorial Aerobic Scope", paste("(MMR/SMR)"))),
       x = "Temperature (°C)") +
  scale_color_manual(values = c("lightslategray", "lightblue", "lightcoral", "indianred4")) +
  theme_classic() +
  theme(axis.text = element_text(color = "black", size = 12),
        axis.title = element_text(size = 13),
        panel.border = element_rect(colour = "black", fill = NA, size = 0.7),
        plot.caption = element_text(size = 10, hjust = 0),
        legend.position = "none")
```

####**Part 3.** Plot resting heart rate of lobsters in each temperature treatment 
```{r}
##############################
# add dummy data for 26C
##############################

dummy_data <- c("NA", "26", "NA", "NA", "NA")

# combine data with dummy data
resting_heart_rates2 <- rbind(resting_heart_rates, dummy_data) %>% 
  mutate(rhr = as.numeric(as.character(rhr)))

##############################
# rhr panel
##############################

rhr_plot <- ggplot() +
  geom_jitter(data = resting_heart_rates2, mapping = aes(x = as.factor(temp), y = rhr), alpha = 0.2, width = 0.2, size = 3) +
  geom_errorbar(data = rhr_sum, aes(x = as.factor(temp), ymin = hr_avg - error, ymax = hr_avg + error), width = 0.1, size = 1, color = "black") +
  geom_point(data = rhr_sum, aes(x = as.factor(temp),  y = hr_avg), size = 6, color = "black") +
  geom_point(data = rhr_sum, aes(x = as.factor(temp),  y = hr_avg, color = as.factor(temp)), size = 4.5) +
  labs(y = expression(atop("Resting Heart Rate", paste("(beats",~min^-1,")"))),
       x = "Temperature (°C)") +
  scale_color_manual(values = c("lightslategray", "lightblue", "lightcoral")) +
  theme_classic() +
  theme(axis.text = element_text(color = "black", size = 12),
        axis.title = element_text(size = 13),
        panel.border = element_rect(colour = "black", fill = NA, size = 0.7),
        plot.caption = element_text(size = 10, hjust = 0),
        legend.position = "none")
```

####**Part 4.** Combine panels and print plot
```{r}
##############################
# combine panels and save plot
##############################

mr_hr_plot <- cowplot::plot_grid(smr_plot, mmr_plot, aas_plot, fas_plot, rhr_plot, nrow = 3, align = "vh", labels = "AUTO")

cowplot::save_plot(here::here("figures", "main_text", "Fig2.pdf"), mr_hr_plot, base_width = 8, base_height = 9)
```
