==> code/07_EPOC_data_processing.Rmd <==
---
title: "EPOC Data Processing"
author: "Sam Csik"
date: "7/6/2020"
output: html_document
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

###**Summary**

The metabolic cost of recovery is termed excess post-exercise oxygen consumption (EPOC) and is measured by esitmating the total amount of oxygen consumed during the recovery period ([Zhang et al. 2018](https://onlinelibrary.wiley.com/doi/pdf/10.1111/jfb.13593?casa_token=JnP8Y3k3fPoAAAAA:FsQV5g-1WeOs-5x0mpFZSamYvDwZz51OiJeC5nQsHZulAbVH4TaQlBOeplnxw3sZjbrAX3oTv5i2ap8)).

###**Outline**

**Part 0.** Import libraries and data
**Part 1.** Total EPOC summary stats 
**Part 2.** Hourly EPOC summary stats
**Part 3.** Time to EPOC summary stats
**Part 4.** EPOC at MMR~50~ summary stats
**Part 5.** Time to MMR~50~ summary stats
**Part 6.** MMR~50~ summary stats

###**Packages Required:** 

    - tidyverse
    - here

###**Data Required:**

**(file path: data/metabolism)**

    - epoc.csv

####**Part 0.** Import packages and data  
```{r}
##############################
# import libraries
##############################

source(here::here("code", "00_libraries.R"))

##############################
# load data
##############################

epoc <- read_csv(here::here("data", "metabolism", "epoc.csv"))
```

####**Part 1.** Wrangle total EPOC for each treatment
```{r}
##############################
# total epoc summary
##############################

EPOC_full_summary <- epoc %>% 
  filter(spar == 0.2) %>% 
  group_by(temp) %>% 
  summarize(
    epoc_full_avg = mean(EPOC_full, na.rm = T),
    epoc_full_sd = sd(EPOC_full, na.rm = T),
    epoc_full_se = epoc_full_sd/sqrt(length(EPOC_full)),
    error = qnorm(0.975) * sd(EPOC_full)/sqrt(length(EPOC_full)),
    epoc_full_upper95 = epoc_full_avg + error,
    epoc_full_lower95 = epoc_full_avg - error
  )

# write.csv(EPOC_full_summary, here::here("data", "metabolism", "outputs", "epoc_full_summary_stats.csv"), row.names = FALSE)
```

####**Part 2.** Wrangle hourly EPOC for each treatment
```{r}
##############################
# hourly epoc summary
##############################

EPOC_hourly_summary <- epoc %>% 
  filter(spar == 0.2) %>% 
  select(ID, temp, EPOC_1hr, EPOC_2hr, EPOC_3hr, EPOC_4hr, EPOC_5hr) %>% 
  mutate(`0-1` = EPOC_1hr, # currently cumulative; need to subtract 
         `1-2` = EPOC_2hr - EPOC_1hr,
         `2-3` = EPOC_3hr - EPOC_2hr,
         `3-4` = EPOC_4hr - EPOC_3hr,
         `4-5` = EPOC_5hr - EPOC_4hr) %>% 
  select(ID, temp, `0-1`, `1-2`, `2-3`, `3-4`, `4-5`) %>% 
  mutate(temp = as.factor(temp)) %>% 
  group_by(temp) %>% 
  summarize(
    #### averages ####
    avg01 = round(mean(`0-1`), 3),
    avg12 = round(mean(`1-2`), 3),
    avg23 = round(mean(`2-3`), 3),
    avg34 = round(mean(`3-4`), 3),
    avg45 = round(mean(`4-5`), 3),
    #### standard deviations ####
    avg01_sd = round(sd(`0-1`, na.rm = T), 3),
    avg12_sd = round(sd(`1-2`, na.rm = T), 3),
    avg23_sd = round(sd(`2-3`, na.rm = T), 3),
    avg34_sd = round(sd(`3-4`, na.rm = T), 3),
    avg45_sd = round(sd(`4-5`, na.rm = T), 3),
    #### standard error ####
    avg01_se = avg01_sd/sqrt(length(avg01)),
    avg12_se = avg12_sd/sqrt(length(avg12)),
    avg23_se = avg23_sd/sqrt(length(avg23)),
    avg34_se = avg34_sd/sqrt(length(avg34)),
    avg45_se = avg45_sd/sqrt(length(avg45)),
    #### error ####
    avg01_error = qnorm(0.975) * sd(`0-1`)/sqrt(length(`0-1`)),
    avg12_error = qnorm(0.975) * sd(`1-2`)/sqrt(length(`1-2`)),
    avg23_error = qnorm(0.975) * sd(`2-3`)/sqrt(length(`2-3`)),
    avg34_error = qnorm(0.975) * sd(`3-4`)/sqrt(length(`3-4`)),
    avg45_error = qnorm(0.975) * sd(`4-5`)/sqrt(length(`4-5`)),
    #### uppper ci ####
    upper_95ci_01 = avg01 + avg01_error,
    upper_95ci_12 = avg12 + avg12_error,
    upper_95ci_23 = avg23 + avg23_error,
    upper_95ci_34 = avg34 + avg34_error,
    upper_95ci_45 = avg45 + avg45_error,
    #### lower ci ####
    lower_95ci_01 = avg01 - avg01_error,
    lower_95ci_12 = avg12 - avg12_error,
    lower_95ci_23 = avg23 - avg23_error,
    lower_95ci_34 = avg34 - avg34_error,
    lower_95ci_45 = avg45 - avg45_error)

##############################
# convert each summary stat to long format and combine into single df
##############################

EPOC_treatment_averages <- EPOC_hourly_summary %>% 
  select(temp, avg01, avg12, avg23, avg34, avg45) %>% 
  pivot_longer(cols = c(avg01, avg12, avg23, avg34, avg45), names_to = "hour_label1", values_to = "mean")

EPOC_treatment_sd <- EPOC_hourly_summary %>% 
  select(avg01_sd, avg12_sd, avg23_sd, avg34_sd, avg45_sd) %>% 
  pivot_longer(cols = c(avg01_sd, avg12_sd, avg23_sd, avg34_sd, avg45_sd), names_to = "hour_label2", values_to = "sd")

EPOC_treatment_se <- EPOC_hourly_summary %>% 
  select(avg01_se, avg12_se, avg23_se, avg34_se, avg45_se) %>% 
  pivot_longer(cols = c(avg01_se, avg12_se, avg23_se, avg34_se, avg45_se), names_to = "hour_label3", values_to = "se")

EPOC_treatment_upperci <- EPOC_hourly_summary %>% 
  select(upper_95ci_01, upper_95ci_12, upper_95ci_23, upper_95ci_34, upper_95ci_45) %>% 
  pivot_longer(cols = c(upper_95ci_01, upper_95ci_12, upper_95ci_23, upper_95ci_34, upper_95ci_45), names_to = "hour_label4", values_to = "upper_95ci")

EPOC_treatment_lowerci <- EPOC_hourly_summary %>% 
  select(lower_95ci_01, lower_95ci_12, lower_95ci_23, lower_95ci_34, lower_95ci_45) %>% 
  pivot_longer(cols = c(lower_95ci_01, lower_95ci_12, lower_95ci_23, lower_95ci_34, lower_95ci_45), names_to = "hour_label5", values_to = "lower_95ci")

##############################
# combine dfs
##############################

EPOC_hourly_wrangled <- cbind(EPOC_treatment_averages, EPOC_treatment_sd, 
                              EPOC_treatment_se, EPOC_treatment_lowerci, 
                              EPOC_treatment_upperci) %>% 
  select(temp, hour_label1, mean, sd, se, lower_95ci, upper_95ci) %>% 
  mutate(
    hour = case_when(
      hour_label1 == "avg01" ~ "0-1",
      hour_label1 == "avg12" ~ "1-2",
      hour_label1 == "avg23" ~ "2-3",
      hour_label1 == "avg34" ~ "3-4",
      hour_label1 == "avg45" ~ "4-5"
    )
  )

# write.csv(EPOC_hourly_wrangled, here::here("data", "metabolism", "outputs", "epoc_hourly_summary_stats.csv"), row.names = FALSE)
```

####**Part 3.** Wrangle time to EPOC data
```{r}
time_to_EPOC_full_summary <- epoc %>% 
  filter(spar == 0.2) %>% 
  group_by(temp) %>% 
  summarize(
    time_to_epoc_avg = round(mean(end_EPOC_min), 3),
    time_to_epoc_sd = round(sd(end_EPOC_min, na.rm = T), 3),
    time_to_epoc_se = time_to_epoc_sd/sqrt(length(end_EPOC_min)),
    error = qnorm(0.975) * sd(end_EPOC_min)/sqrt(length(end_EPOC_min)),
    time_to_epoc_upper95 = time_to_epoc_avg + error,
    time_to_epoc_lower95 = time_to_epoc_avg - error
  )

# write.csv(time_to_EPOC_full_summary, here::here("data", "metabolism", "outputs", "time_to_epoc_full_summary_stats.csv"), row.names = FALSE)
```

####**Part 4.** Wrangle EPOC at MMR~50~
```{r}
EPOC_at_MMR50_summary <- epoc %>% 
  filter(spar == 0.2) %>% 
  group_by(temp) %>% 
  summarize(
    epoc_at_MMR50_avg = round(mean(EPOC_mmr), 3),
    epoc_at_MMR50_sd = round(sd(EPOC_mmr, na.rm = T), 3),
    epoc_at_MMR50_se = epoc_at_MMR50_sd/sqrt(length(EPOC_mmr)),
    error = qnorm(0.975) * sd(EPOC_mmr)/sqrt(length(EPOC_mmr)),
    epoc_at_MMR50_upper95 = epoc_at_MMR50_avg + error,
    epoc_at_MMR50_lower95 = epoc_at_MMR50_avg - error
  )

# write.csv(EPOC_at_MMR50_summary, here::here("data", "metabolism", "outputs", "epoc_at_MMR50_summary_stats.csv"), row.names = FALSE)
```

####**Part 5.** Wrangle time to MMR~50~
```{r}
time_to_MMR50_summary <- epoc %>% 
  filter(spar == 0.2) %>% 
  group_by(temp) %>% 
  summarize(
    time_to_MMR50_avg = round(mean(end_EPOC_mmr), 3),
    time_to_MMR50_sd = round(sd(end_EPOC_mmr, na.rm = T), 3),
    time_to_MMR50_se = time_to_MMR50_sd/sqrt(length(end_EPOC_mmr)),
    error = qnorm(0.975) * sd(end_EPOC_mmr)/sqrt(length(end_EPOC_mmr)),
    time_to_MMR50_upper95 = time_to_MMR50_avg + error,
    time_to_MMR50_lower95 = time_to_MMR50_avg - error
  )

# write.csv(time_to_MMR50_summary, here::here("data", "metabolism", "outputs", "time_to_MMR50_summary_stats.csv"), row.names = FALSE)
```

####**Part 6.** Wrangle MMR~50~
```{r}
MMR50_summary <- epoc %>% 
  filter(spar == 0.2) %>% 
  group_by(temp) %>% 
  summarize(
    MMR50_avg = round(mean(MO2_mmr), 3),
    MMR50_sd = round(sd(MO2_mmr), 3),
    MMR50_se = MMR50_sd/sqrt(length(MO2_mmr)),
    error = qnorm(0.975) * sd(MO2_mmr)/sqrt(length(MO2_mmr)),
    MMR50_upper95 = MMR50_avg + error,
    MMR50_lower95 = MMR50_avg - error
  )

# write.csv(MMR50_summary, here::here("data", "metabolism", "outputs", "MMR50_summary_stats.csv"), row.names = FALSE)
```
