==> code/10_max_consumption_analysis&plotting.Rmd <==
---
title: "Max Consumption Analysis"
author: "Sam Csik"
date: "6/24/2020"
output: html_document
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

###**Summary**

We examine the relationship between lobster maximum consumption rates and temperature by fitting a linear mixed effects model to the data.

###**Outline**

**Part 0** Import packages and data
**Part 1** (TABLE S9 GENERATED HERE) Fit model to data
**Part 2** (FIGURE 3 GENERATED HERE) Plot maximum consumption rates across temperatures

###**Required Packages:**

    - tidyverse
    - lme4

###**Required Data:**

**(file path: data/foraging/raw)**

    - foraging_assay_data.csv 
    - max_consump_summary_stats.csv

####**Part 0.** Import libraries and data
```{r}
##############################
# load required packages
##############################

source(here::here("code", "00_libraries.R"))

##############################
# load data
##############################

foraging_assays <- read_csv(here::here("data", "foraging", "raw", "foraging_assay_data.csv")) 

mc_summary <- read_csv(here::here("data", "foraging", "outputs", "max_consump_summary_stats.csv"))
```   

####**Part 1.** Fit nonlinear mixed effects model to *ad lib* foraging trial data (60 mussels only) and generate table S9
```{r}
##############################
# filter data for only trials with 60 mussels
##############################

foraging60 <- foraging_assays %>% 
  filter(Initial == "60")

##############################
# fit lmer (3 levels (assays) per lobster, but need to account for lobster as random effect so still using this method)
##############################

consump_mod <- lmerTest::lmer(Killed ~ temp + (1|lobster_id), foraging60) # varying slope only
summary(consump_mod)
summary(multcomp::glht(consump_mod, linfct = multcomp::mcp(temp = "Tukey")))

#analyze as continuous?
foraging60b<-foraging60 %>%
    mutate(temp = as.numeric(temp))

cmod<-lmerTest::lmer(Killed ~ temp + (1|lobster_id), foraging60b)
  summary(cmod)


##############################
# diagnostics
##############################

plot(consump_mod) # no obvious deviations from linear form; relatively constant variance for all data (good!)

qqmath(consump_mod, id = 0.05) # id identifies values that may be exerting undue influence on the model (i.e. outliers); fairly normal (good)

##############################
# create table (manually edit aesthetics in Word)
##############################

tab_model(consump_mod, 
          dv.labels = "mussels consumed",
          pred.labels = c(`(Intercept)` = "Intercept",
                          `temp [16]` = "16°C",
                          `temp [21]` = "21°C",
                          `temp [26]` = "26°C"),
          string.pred = "Coefficient",
          string.est = "Estimate",
          show.se = TRUE,
          string.se = "SE",
          string.ci = "95% CI")
```

####**Part 2.** Plot maximum consumption rates
```{r}
##############################
# filter for only ad lib feeding assays
##############################

foraging60 <- foraging_assays %>% 
  filter(Initial == "60")

##############################
# plot
##############################

consumption_plot <- ggplot() + 
  geom_jitter(data = foraging60, aes(x = temp, y = Killed), color = "black", width = 0.2, alpha = 0.2, size = 3) + 
  geom_smooth(data = foraging60,aes(x=temp,y=Killed),colour = "black", method = "lm",se=F)+
  geom_errorbar(data = mc_summary, aes(x = temp, ymin = consump_avg - error, ymax = consump_avg + error), width = 0.1, size = 1, color = "black") +
  geom_point(data = mc_summary, aes(x =  temp,  y = consump_avg), size = 6, colour = "black") +
  geom_point(data = mc_summary, aes(x = temp,  y = consump_avg, color = as.factor(temp)), size = 4.5) +
  labs(y = expression(atop("Maximum Consumption Rate", paste("(prey consumed" ~ predator^{-1} ~ "24",~hr^{-1},")"))),
       x = expression(paste("Temperature (", degree, "C)"))) +
  ylim(0, 35) +
  scale_color_manual(values=c("lightslategray", "lightblue", "lightcoral", "indianred4")) +
  theme_classic() + 
  theme(axis.text = element_text(color = "black", size = 12),
        axis.title = element_text(size = 13),
        panel.border = element_rect(colour = "black", fill = NA, size = 0.7),
        plot.caption = element_text(size = 10, hjust = 0),
        legend.position = "none")+
      scale_x_continuous(limits=c(10, 28),breaks=c(11,16,21,26)) 


cowplot::save_plot(here::here("figures", "main_text", "Fig3.pdf"), consumption_plot, base_width = 5, base_height = 4)
```
