==> code/12_activation_energy_analysis.Rmd <==
---
title: "Estimating Activation Energies"
author: "Sam Csik"
date: "6/25/2020"
output: html_document
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

###**Summary**

I apply the logic in [Englund et al. 2011](https://onlinelibrary.wiley.com/doi/epdf/10.1111/j.1461-0248.2011.01661.x) for modeling attack rates and max intake rates as a function of temperature to my metabolism and foraging data data by plotting ln(rate) ~ -1/kT and fitting a linear model to estimate the activation energy , E~a~, for each rate.

According to Englund et al. 2011, a and I~max~ should be functions of the Arrhenius temperature, -1/(kT) where k is Boltzmann's constant (eV K^-1^) and T is temperature (Kelvin).

You can then fit the Arrhenius equation: 

Y = c*e*^E~a~(-1/kT)^ to the data where Y is a reaction rate, c is a fitted constant, and E~a~ is the fitted activation energy (eV) that describes the slope of the response.

I estimate E~a~ using the linearized form of the equation: 

>log(y) = log(c) + E~a~(-1/kT)

MTE predicts that the temperature dependence of SMR and foraging rates should be constrained between E~a~ = 0.6-0.7. 

NOTE: Because many foraging assay trials for 11C-acclimated lobsters resulted in little to no consumption, we omitted this temperature treatement from E~a~ analyses.

###**Outline**

**Part 0** Import packages and data
**Part 1** Data wrangling to add Arrhenius temperature to dfs
**Part 2** (FIG S3 GENERATED HERE) Create Arrhenius plots for 16, 21, 26C treatments.

###**Required Packages:**

    - tidyverse
    - here

###**Required Data:**

**(file path: data/metabolism/)**

    - metabolic_traits.csv

**(file path: data/foraging/raw)**

    - foraging_assay_data.csv
    
**(file path: data/foraging/outputs)**

    - ind_FR_estimates.csv

####**Part 0.** Import packages and data; also define Boltzmann's constant
```{r}
##############################
# import packages
##############################

source(here::here("code", "00_libraries.R"))

##############################
# import data
##############################

metabolic_rates <- read_csv(here::here("data", "metabolism", "metabolic_traits.csv")) 

foraging_assays <- read_csv(here::here("data", "foraging", "raw", "foraging_assay_data.csv")) 

functional_response_params <- read_csv(here::here("data", "foraging", "outputs", "ind_FR_estimates.csv"))

##############################
# Boltzmann's constant (k)
##############################

k = 8.617e-5 # units: eV K^-1
```

####**Part 1.** Add Arrhenius temperature add data frames
```{r}
##############################
# metabolic rates
##############################

Ea_metabolic_rates <- metabolic_rates %>% 
  filter(temp != "11") %>% 
  mutate(Kelvin = temp + 273.15, 
         Arrhenius_temp = -1/(k*Kelvin))

##############################
# foraging assay data (for max consumption, use only trials with 60 mussels)
##############################

Ea_foraging_assays <- foraging_assays %>% 
  filter(Initial == "60",
         temp != "11") %>% 
  mutate(Kelvin = temp + 273.15,
         Arrhenius_temp = -1/(k*Kelvin))

##############################
# functional response parameters
##############################

Ea_functional_response_params <- functional_response_params %>% 
  filter(temp != "11") %>% 
  mutate(Kelvin = temp + 273.15,
         Arrhenius_temp = -1/(k*Kelvin))

# subset a and h separately 
Ea_a <- Ea_functional_response_params %>% filter(parameter == "a")
Ea_h <- Ea_functional_response_params %>% filter(parameter == "h")
```

####**Part 2.** Fit the linearized form of the Arrhenius equation to the data (on a log scale) to estimate Ea for each rate
```{r}
##############################
# smr 
##############################

Ea_smr_mod <- lm(log(SMR) ~ Arrhenius_temp, Ea_metabolic_rates)
summary(Ea_smr_mod)

##############################
# mmr
##############################

Ea_mmr_mod <- lm(log(MMR) ~ Arrhenius_temp, Ea_metabolic_rates)
summary(Ea_mmr_mod)

##############################
# aas
##############################

Ea_aas_mod <- lm(log(AAS) ~ Arrhenius_temp, Ea_metabolic_rates)
summary(Ea_aas_mod)

##############################
# max consumption
##############################

Ea_max_consump_mod <- lm(log(Killed) ~ Arrhenius_temp, Ea_foraging_assays)
summary(Ea_max_consump_mod)

##############################
# attack rate
##############################

Ea_a_mod <- lm(log(estimate) ~ Arrhenius_temp, Ea_a)
summary(Ea_a_mod)

##############################
# handling time
##############################

Ea_h_mod <- lm(log(estimate) ~ Arrhenius_temp, Ea_h)
summary(Ea_h_mod)
```

####**Part 2.** Plot each ln(rate) ~ -1/kT
```{r}
##############################
# smr 
##############################

Ea_smr_plot <- ggplot(Ea_metabolic_rates, aes(x = Arrhenius_temp, y = log(SMR))) +
  geom_point(aes(color = as.factor(temp), shape = as.factor(temp), size = as.factor(temp)), alpha = 0.8, name = "Temperature (°C)") +
  scale_shape_manual(values = c(16, 17, 18), name = "Temperature (°C)") +
  scale_size_manual(values = c(4, 4, 5), name = "Temperature (°C)") +
  stat_smooth(method = "lm", se = FALSE, color = "black") + # this matches model fit
  xlab("Temperature (-1/kT)") +
  ylab(expression("ln(SMR)" ~ "(mg" ~ O[2] ~ kg^{-1} ~ min^{-1},")")) +
  scale_color_manual(values = c("lightblue", "lightcoral", "indianred4")) +
  theme_classic() +
  theme(axis.text = element_text(color = "black", size = 12),
        axis.title = element_text(size = 13),
        panel.border = element_rect(colour = "black", fil = NA, size = 0.7),
        legend.position = "none") 

##############################
# mmr
##############################

Ea_mmr_plot <- ggplot(Ea_metabolic_rates, aes(x = Arrhenius_temp, y = log(MMR))) +
  geom_point(aes(color = as.factor(temp), shape = as.factor(temp), size = as.factor(temp)), alpha = 0.8, name = "Temperature (°C)") +
  scale_shape_manual(values = c(16, 17, 18), name = "Temperature (°C)") +
  scale_size_manual(values = c(4, 4, 5), name = "Temperature (°C)") +
  stat_smooth(method = "lm", se = FALSE, color = "black") + # this matches model fit
  xlab("Temperature (-1/kT)") +
  ylab(expression("ln(MMR)" ~ "(mg" ~ O[2] ~ kg^{-1} ~ min^{-1} ~ ")")) +
  scale_color_manual(values = c("lightblue", "lightcoral", "indianred4")) +
  theme_classic() +
  theme(axis.text = element_text(color = "black", size = 12),
        axis.title = element_text(size = 13),
        panel.border = element_rect(colour = "black", fil = NA, size = 0.7),
        legend.position = "none")

##############################
# aas
##############################

Ea_aas_plot <- ggplot(Ea_metabolic_rates, aes(x = Arrhenius_temp, y = log(AAS))) +
  geom_point(aes(color = as.factor(temp), shape = as.factor(temp), size = as.factor(temp)), alpha = 0.8, name = "Temperature (°C)") +
  scale_shape_manual(values = c(16, 17, 18), name = "Temperature (°C)") +
  scale_size_manual(values = c(4, 4, 5), name = "Temperature (°C)") +
  stat_smooth(method = "lm", se = FALSE, color = "black") + # this matches model fit
  xlab("Temperature (-1/kT)") +
  ylab(expression("ln(AS)" ~ "(mg" ~ O[2] ~ kg^{-1} ~ min^{-1} ~ ")")) +
  scale_color_manual(values = c("lightblue", "lightcoral", "indianred4")) +
  theme_classic() +
  theme(axis.text = element_text(color = "black", size = 12),
        axis.title = element_text(size = 13),
        panel.border = element_rect(colour = "black", fil = NA, size = 0.7),
        legend.position = "none")

##############################
# max consumption
##############################

Ea_max_consump_plot <- ggplot(Ea_foraging_assays, aes(x = Arrhenius_temp, y = log(Killed))) +
  geom_point(aes(color = as.factor(temp), shape = as.factor(temp), size = as.factor(temp)), alpha = 0.8, name = "Temperature (°C)") +
  scale_shape_manual(values = c(16, 17, 18), name = "Temperature (°C)") +
  scale_size_manual(values = c(4, 4, 5), name = "Temperature (°C)") +
  stat_smooth(method = "lm", se = FALSE, color = "black") + # this matches model fit
  xlab("Temperature (-1/kT)") +
  ylab(expression("ln(Mussels Consumed)")) +
  scale_color_manual(values = c("lightblue", "lightcoral", "indianred4")) +
  theme_classic() +
  theme(axis.text = element_text(color = "black", size = 12),
        axis.title = element_text(size = 13),
        panel.border = element_rect(colour = "black", fil = NA, size = 0.7),
        legend.position = "none") 

##############################
# attack rate
##############################

Ea_a_plot <- ggplot(Ea_a, aes(x = Arrhenius_temp, y = log(estimate))) + 
  geom_point(aes(color = as.factor(temp), shape = as.factor(temp), size = as.factor(temp)), alpha = 0.8,  name = "Temperature (°C)") +
  scale_shape_manual(values = c(16, 17, 18), name = "Temperature (°C)") +
  scale_size_manual(values = c(4, 4, 5), name = "Temperature (°C)") +
  stat_smooth(method = "lm", se = FALSE, color = "black") + # this matches model fit
  xlab("Temperature (-1/kT)") +
  ylab(expression("ln(attack rate)")) +
  scale_color_manual(values = c("lightblue", "lightcoral", "indianred4")) +
  theme_classic() +
  theme(axis.text = element_text(color = "black", size = 12),
        axis.title = element_text(size = 13),
        panel.border = element_rect(colour = "black", fil = NA, size = 0.7),
        legend.position = "none") 

##############################
# handling time
##############################

Ea_h_plot <- ggplot(Ea_h, aes(x = Arrhenius_temp, y = log(estimate))) +
  geom_point(aes(color = as.factor(temp), shape = as.factor(temp), size = as.factor(temp)), alpha = 0.8, name = "Temperature (°C)") +
  scale_shape_manual(values = c(16, 17, 18), name = "Temperature (°C)") +
  scale_size_manual(values = c(4, 4, 5), name = "Temperature (°C)") +
  stat_smooth(method = "lm", se = FALSE, color = "black") + # this matches model fit
  xlab("Temperature (-1/kT)") +
  ylab(expression("ln(handling time)")) +
  scale_color_manual(values = c("lightblue", "lightcoral", "indianred4")) +
  theme_classic() +
  theme(axis.text = element_text(color = "black", size = 12),
        axis.title = element_text(size = 13),
        panel.border = element_rect(colour = "black", fil = NA, size = 0.7),
        legend.position = "none") 

##############################
# combine and print plots
##############################

Ea_plots_no11 <- cowplot::plot_grid(Ea_smr_plot, Ea_mmr_plot, Ea_aas_plot, Ea_a_plot, Ea_h_plot, Ea_max_consump_plot, nrow = 2, align = "v", labels = "AUTO")

# cowplot::save_plot(here::here("figures", "FigS3.pdf"), Ea_plots_no11, base_width = 15, base_height = 10) 
```
