==> code/13_correlations_data_processing.Rmd <==
---
title: "Correlational Analysis Between Lobster Metabolism and Foraging - Data Processing"
author: "Sam Csik"
date: "6/25/2020"
output: html_document
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

###**Summary**

Combine metabolism and foraging data into the same data frame for correlational analysis and plotting.

###**Outline**

**Part 0** Import packages and data
**Part 1** Wrangle and combine metabolism and foraging data

###**Required Packages**

    - tidyverse
    - here
    
###**Required Data**

**(file path: data/metabolism/)**

    - metabolic_traits.csv
    - epoc.csv

**(file path: data/foraging/raw)**

    - foraging_assay_data.csv
    
####**Part 0.** Import packages and data
```{r}
##############################
# load required packages
##############################

source(here::here("code", "00_libraries.R"))

##############################
# load data
##############################

metabolic_rates <- read_csv(here::here("data", "metabolism", "metabolic_traits.csv"))

foraging_assays <- read_csv(here::here("data", "foraging", "raw", "foraging_assay_data.csv")) 
```

####**Part 1.** Get both metabolism and foraging data into the same format; combine and print
```{r}
##############################
# metabolic rates
##############################

NEW_metabolic_rates <- metabolic_rates %>% 
  dplyr::rename(lobster_id = ID) %>% 
  select(lobster_id, temp, SMR, MMR, AAS, FAS) %>% 
  mutate(temp = as.character(temp)) 

##############################
# foraging assays
##############################

NEW_foraging_assays <- foraging_assays %>% 
  filter(Initial == "60") %>% 
  dplyr::group_by(lobster_id) %>% 
  dplyr::summarize(avg_eaten = mean(Killed),
            sd_eaten = sd(Killed),
            se_eaten = sd(Killed/sqrt(length(Killed))),
            error = qnorm(0.975) * sd(Killed)/sqrt(length(Killed)),
            upper_95ci = avg_eaten + error,
            lower_95ci = avg_eaten - error) %>% 
  mutate(avg_eaten = as.numeric(avg_eaten))

##############################
# combine data sets and print
##############################

metabolism_foraging_data <- full_join(NEW_metabolic_rates, NEW_foraging_assays)

# write.csv(metabolism_foraging_data, here::here("data", "correlations", "outputs", "metabolism_foraging_data.csv"), row.names = FALSE)
```
