==> code/14_correlations_analysis&plotting.Rmd <==
---
title: "Correlational Analysis Between Lobster Metabolism and Foraging - Analysis & Plotting"
author: "Sam Csik"
date: "6/25/2020"
output: html_document
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

###**Summary**

To determine how different metabolic traits alter lobster foraging across temperatures, we correlate SMR, MMR, AAS, and FAS with max consumption rates. 

###**Outline**

**Part 0** Import packages and data
**Part 1** Test Pearsons's r assumptions
**Part 2** Compute r and p-values
**Part 3** (TABLE 1 GENERATED HERE) Create table for correlational analysis
**Part 4** (FIGURE S4 GENERATED HERE) Build correlation matrix
**Part 5** (FIGURE 5 GENERATED HERE) Plot correlations between each metabolic trait and max consumption rates

###**Required Packages:**

    - tidyverse
    - here
    - cowplot
    - ggExtra
    - ggpubr

###**Required Data:**

**(file path: data/correlations/outputs/)**

    - metabolism_foraging_data.csv

####**Part 0.** Load packages and data   
```{r}
##############################
# import packages
##############################

source(here::here("code", "00_libraries.R"))

##############################
# load data
##############################

metabolism_foraging_data <- read_csv(here::here("data", "correlations", "outputs", "metabolism_foraging_data.csv")) %>% 
  mutate(temp = as.factor(temp))
```

####**Part 1.** Pearson's correlation (r) measures a linear dependence betwene two variables (x and y). It can only be used when x and y are normally distributed. First, check the following assumptions:

1. Is the covariation linear? (for our data, I would argue yes)
2. Are the data normally distributed (use Shapiro-Wilk normality test and normality plots) (test below)
```{r}
##############################
# Shapiro test for each metabolic trait and avg eaten 
  # N0 = data are normal
  # NA = data are not normal
##############################

shapiro.test(metabolism_foraging_data$SMR) # p = 0.1281
shapiro.test(metabolism_foraging_data$MMR) # p = 0.8752
shapiro.test(metabolism_foraging_data$AAS) # p = 0.9025
shapiro.test(metabolism_foraging_data$FAS) # p = 0.001497* bc of 11C
shapiro.test(metabolism_foraging_data$avg_eaten) # p = 0.2052

##############################
# visual inspection using Q-Q plots
##############################

ggqqplot(metabolism_foraging_data$SMR, ylab = "SMR") # pretty good
ggqqplot(metabolism_foraging_data$MMR, ylab = "MMR") # great
ggqqplot(metabolism_foraging_data$AAS, ylab = "AAS") # great
ggqqplot(metabolism_foraging_data$FAS, ylab = "FAS") # great except for 11C
ggqqplot(metabolism_foraging_data$avg_eaten, ylab = "Average Consumption") # ok
```

**Part 2.** Compute the correlation coefficient and significance level for each metabolic trait; extract p-values and coefficients to build Table 1
```{r}
##############################
# smr
##############################

smr_cor <- cor.test(metabolism_foraging_data$SMR, metabolism_foraging_data$avg_eaten, method = c("pearson"))

smr_cor_p <- round(smr_cor$p.value, 3)
smr_cor_est <- round(smr_cor$estimate, 3)

##############################
# mmr
##############################

mmr_cor <- cor.test(metabolism_foraging_data$MMR, metabolism_foraging_data$avg_eaten, method = c("pearson"))

mmr_cor_p <- round(mmr_cor$p.value, 3)
mmr_cor_est <- round(mmr_cor$estimate, 3)

##############################
# aas
##############################

aas_cor <- cor.test(metabolism_foraging_data$AAS, metabolism_foraging_data$avg_eaten, method = c("pearson"))

aas_cor_p <- round(aas_cor$p.value, 3)
aas_cor_est <- round(aas_cor$estimate, 3)

##############################
# fas
##############################

fas_cor <- cor.test(metabolism_foraging_data$FAS, metabolism_foraging_data$avg_eaten, method = c("pearson"))

fas_cor_p <- round(fas_cor$p.value, 3)
fas_cor_est <- round(fas_cor$estimate, 3)
```

####**Part 3.** Build Table 1
```{r}
##############################
# build table
##############################

col_names <- c("Metabolic Trait", "SMR", "MMR", "AAS", "FAS")
cor_estimates <- c(smr_cor_est, mmr_cor_est, aas_cor_est, fas_cor_est)
cor_pvals <- c(smr_cor_p, mmr_cor_p, aas_cor_p, fas_cor_p)

correlation_table <- data.frame("Metabolic Trait" = c("SMR", "MMR", "AAS", "FAS"),
                                "Pearson's Correlation Coefficient (r)" = c(smr_cor_est, mmr_cor_est, aas_cor_est, fas_cor_est),
                                "R2" = c(smr_cor_est^2, mmr_cor_est^2, aas_cor_est^2, fas_cor_est^2),
                                "p-value" = c(smr_cor_p, mmr_cor_p, aas_cor_p, fas_cor_p))

# write.csv(correlation_table, here::here("data", "correlations", "outputs", "correlations.csv"), row.names = FALSE)
```

####**Part 4.** Create correlation matrix plot (Fig S4)
```{r}
##############################
# simplify df
##############################

# rename variables in metabolism_foraging_data for ease of viewing
NEW_metabolism_foraging_data <- metabolism_foraging_data %>% 
  select(SMR, MMR, AAS, FAS, avg_eaten) %>% 
  rename(`Average Consumed` = avg_eaten)

##############################
# calculate pairwise Pearson's r for correlations
##############################

corr_df <- cor(NEW_metabolism_foraging_data)

##############################
# construct correlation matrix 
##############################

# pdf(here::here("figures","FigS4.pdf"))
corr_matrix <- corrplot(corr_df,
         type = "upper",
         method = "ellipse", 
         tl.col = "black", 
         tl.cex = 0.5)
# dev.off()
```

####**Part 5.** Build plots correlating each metabolic trait and max consumption rate (averaged across three trials for each lobster; Fig 5)
```{r}
##############################
# smr
##############################

smr_corr_plot <- ggplot(metabolism_foraging_data, aes(x = SMR, y = avg_eaten)) + 
  geom_errorbar(aes(ymin = avg_eaten - sd_eaten, ymax = avg_eaten + sd_eaten), width = 0.02, color = "gray") +
  geom_point(aes(color = temp, shape = temp, size = temp)) + 
  geom_smooth(method = "lm", color = "gray36", size = 0.75, level = 0.95) +
  scale_color_manual(values = c("lightslategray", "lightblue", "lightcoral", "indianred4"), name = "Temperature (°C)", labels = c("11", "16", "21", "26")) +
  scale_shape_manual(values = c(15, 16, 17, 18), name = "Temperature (°C)", labels = c("11", "16", "21", "26")) +
  scale_size_manual(values = c(4, 4, 4, 5), name = "Temperature (°C)", labels = c("11", "16", "21", "26")) +
  labs(x = expression(atop("Standard Metabolic Rate", paste("(",mg~O[2]~kg^-1~min^-1,")"))),
       y = expression(atop("Maximum Consumption Rate", paste("(prey consumed" ~ predator^{-1} ~ "24",~hr^{-1},")")))) +
  scale_x_continuous(breaks = seq(0, 1.5, by = 0.2)) +
  theme_classic() +
  theme(axis.text = element_text(color = "black", size = 12),
        axis.title = element_text(size = 13),
        panel.border = element_rect(colour = "black", fill = NA, size = 0.7),
        plot.caption = element_text(size = 10, hjust = 0),
        legend.title = element_text(size = 8),
        legend.text = element_text(size = 8),
        legend.position = c(0.95, 0.95),
        legend.justification = c("right", "top"),
        legend.box.just = "right",
        legend.box.background = element_rect(color = "black", size = 1.1),
        legend.margin = margin(3, 3, 3, 3)) +
  annotate(geom = "curve", x = 0.3, y = 20, xend = 0.184, yend = 9.43,
           curature = 0.2, arrow = arrow(length = unit(2, "mm"))) +
  annotate(geom = "text", x = 0.3, y = 20.1, label = "IV10", hjust = "left") +
  annotate(geom = "curve", x = 1.2, y = 5, xend = 1.31, yend = 14,
           curvature = 0.3, arrow = arrow(length = unit(3, "mm"))) +
  annotate(geom = "text", x = 1.19, y = 5.25, label = "IV19", hjust = "right")

smr_marg <- ggMarginal(smr_corr_plot, type = "density", size = 4, groupColour = TRUE, groupFill = TRUE)

##############################
# mmr
##############################

mmr_corr_plot <- ggplot(metabolism_foraging_data, aes(x = MMR, y = avg_eaten)) + 
  geom_errorbar(aes(ymin = avg_eaten - sd_eaten, ymax = avg_eaten + sd_eaten), width = 0.07, color = "gray") +
  geom_point(aes(color = temp, shape = temp, size = temp)) + 
  geom_smooth(method = "lm", color = "gray36", size = 0.75, level = 0.95) +
  scale_color_manual(values = c("lightslategray", "lightblue", "lightcoral", "indianred4"), name = "Temperature (°C)") +
  scale_shape_manual(values = c(15, 16, 17, 18)) +
  scale_size_manual(values = c(4, 4, 4, 5)) +
  labs(x = expression(atop("Maximum Metabolic Rate", paste("(",mg~O[2]~kg^-1~min^-1,")"))),
       y = expression(atop("Maximum Consumption Rate", paste("(prey consumed" ~ predator^{-1} ~ "24",~hr^{-1},")")))) +
  scale_x_continuous(breaks = seq(0, 7, by = 1)) +
  theme_classic() +
  theme(axis.text = element_text(color = "black", size = 12),
        axis.title = element_text(size = 13),
        panel.border = element_rect(colour = "black", fill = NA, size = 0.7),
        plot.caption = element_text(size = 10, hjust = 0),
        legend.title = element_text(size = 14),
        legend.text = element_text(size = 12),
        legend.position = "none",
        axis.title.y=element_blank(),
        axis.text.y=element_blank(),
        axis.ticks.y=element_blank()) +
  annotate(geom = "curve", x = 3.10, y = 20, xend = 3.4, yend = 9.43,
           curature = 0.15, arrow = arrow(length = unit(3, "mm"))) +
  annotate(geom = "text", x = 3.10, y = 20.1, label = "IV10", hjust = "left") +
  annotate(geom = "curve", x = 4.38, y = 4, xend = 4.25, yend = 13.9,
           curvature = 0.2, arrow = arrow(length = unit(3, "mm"))) +
  annotate(geom = "text", x = 4.37, y = 4.25, label = "IV19", hjust = "right")

mmr_marg <- ggMarginal(mmr_corr_plot, type = "density", size = 4, groupColour = TRUE, groupFill = TRUE)

##############################
# aas
##############################

aas_corr_plot <- ggplot(metabolism_foraging_data, aes(x = AAS, y = avg_eaten)) + 
  geom_errorbar(aes(ymin = avg_eaten - sd_eaten, ymax = avg_eaten + sd_eaten), width = 0.07, color = "gray") +
  geom_point(aes(color = temp, shape = temp, size = temp)) + 
  geom_smooth(method = "lm", color = "gray36", size = 0.75, level = 0.95) +
  scale_color_manual(values = c("lightslategray", "lightblue", "lightcoral", "indianred4"), name = "Temperature (°C)") +
  scale_shape_manual(values = c(15, 16, 17, 18)) +
  scale_size_manual(values = c(4, 4, 4, 5)) +
  labs(x = expression(atop("Absolute Aerobic Scope", paste("(",mg~O[2]~kg^-1~min^-1,")"))),
       y = expression(atop("Maximum Consumption Rate", paste("(prey consumed" ~ predator^{-1} ~ "24",~hr^{-1},")")))) +
  scale_x_continuous(breaks = seq(0, 7, by = 1)) +
  theme_classic() +
  theme(axis.text = element_text(color = "black", size = 12),
        axis.title = element_text(size = 13),
        panel.border = element_rect(colour = "black", fill = NA, size = 0.7),
        plot.caption = element_text(size = 10, hjust = 0),
        legend.title = element_text(size = 14),
        legend.text = element_text(size = 12),
        legend.position = "none") +
  annotate(geom = "curve", x = 3.5, y = 4.5, xend = 3.23, yend = 9.44,
           curature = 0.2, arrow = arrow(length = unit(3, "mm"))) +
  annotate(geom = "text", x = 3.49, y = 4.51, label = "IV10", hjust = "right") +
  annotate(geom = "curve", x = 2.9, y = 20.2, xend = 2.92, yend = 14.13,
           curvature = 0.3, arrow = arrow(length = unit(3, "mm"))) +
  annotate(geom = "text", x = 2.91, y = 20.25, label = "IV19", hjust = "left")

aas_marg <- ggMarginal(aas_corr_plot, type = "density", size = 4, groupColour = TRUE, groupFill = TRUE)

##############################
# fas
##############################

fas_corr_plot <- ggplot(metabolism_foraging_data, aes(x = FAS, y = avg_eaten)) + 
  geom_errorbar(aes(ymin = avg_eaten - sd_eaten, ymax = avg_eaten + sd_eaten), width = 0.3, color = "gray") +
  geom_point(aes(color = temp, shape = temp, size = temp)) + 
  geom_smooth(method = "lm", color = "gray36", size = 0.75, level = 0.95) +
  scale_color_manual(values = c("lightslategray", "lightblue", "lightcoral", "indianred4"), name = "Temperature (°C)") +
  scale_shape_manual(values = c(15, 16, 17, 18)) +
  scale_size_manual(values = c(4, 4, 4, 5)) +
  labs(x = expression(atop("Factorial Aerobic Scope", paste("(MMR/SMR)"))),
       y = expression(atop("Maximum Consumption Rate", paste("(prey consumed" ~ predator^{-1} ~ "24",~hr^{-1},")")))) +
  xlim(3, 19) +
  theme_classic() +
  theme(axis.text = element_text(color = "black", size = 12),
        axis.title = element_text(size = 13),
        panel.border = element_rect(colour = "black", fill = NA, size = 0.7),
        plot.caption = element_text(size = 10, hjust = 0),
        legend.title = element_text(size = 14),
        legend.text = element_text(size = 12),
        legend.position = "none",
        axis.title.y=element_blank(),
        axis.text.y=element_blank(),
        axis.ticks.y=element_blank()) +
  annotate(geom = "curve", x = 17.5, y = 19, xend = 18.22, yend = 9.44,
           curature = 0.5, arrow = arrow(length = unit(3, "mm"))) +
  annotate(geom = "text", x = 17.53, y = 19.1, label = "IV10", hjust = "left") +
  annotate(geom = "curve", x = 3.9, y = 4.45, xend = 3.2, yend = 14.0, 
           curvature = 0.2, arrow = arrow(length = unit(3, "mm"))) +
  annotate(geom = "text", x = 3.85, y = 4.4, label = "IV19", hjust = "right") 

fas_marg <- ggMarginal(fas_corr_plot, type = "density", size = 4, groupColour = TRUE, groupFill = TRUE)

##############################
# combine plots and print
##############################

correlations_plot <- cowplot::plot_grid(smr_marg, mmr_marg, aas_marg, fas_marg, nrow = 2, labels = "AUTO", axis = "tblr", rel_widths = c(0.95, 0.80, 0.95, 0.80))

cowplot::save_plot(here::here("figures", "main_text", "Fig5_updated2.pdf"), correlations_plot, base_width = 10, base_height = 10)

```
